+++ TT_Toolbox_RT/Behaviors/InputString.cpp
@@ -101,7 +101,7 @@ void RemoveCaret(char *str, int &caret, int len)
 void RemoveAllCaret(char *str)
 {
     int i = 0;
-    int len = strlen(str);
+    int len = (int)strlen(str);
     char *tmp = str;
     while (*tmp)
     {
@@ -192,7 +192,7 @@ CKBOOL End(int currentlength, CKBOOL multi, char *str, int &caret, int len, int 
         return PageDown(currentlength, str, caret, len, max);
 
     memmove(str + caret, str + caret + 1, ptr - (str + caret) - 1);
-    caret = ptr - str - 1;
+    caret = (int)(ptr - str - 1);
     str[caret] = '\b';
 
     return TRUE;
@@ -248,7 +248,7 @@ int InputString(const CKBehaviorContext &behcontext)
     int caret = 0;
     beh->GetLocalParameterValue(LOCAL_POS, &caret);
 
-    int currentlength = strlen(str);
+    int currentlength = (int)strlen(str);
 
     CKInputManager *input = (CKInputManager *)behcontext.Context->GetManagerByGuid(INPUT_MANAGER_GUID);
     if (!input)
@@ -268,7 +268,7 @@ int InputString(const CKBehaviorContext &behcontext)
         beh->SetLocalParameterValue(LOCAL_POS, &caret);
 
         CKParameterOut *pout = beh->GetOutputParameter(POUT_STRING);
-        pout->SetValue(str, strlen(str) + 1);
+        pout->SetValue(str, (int)strlen(str) + 1);
         beh->ActivateInput(IN_OFF, FALSE);
         beh->ActivateOutput(OUT_OFF);
         return CKBR_OK;
@@ -284,9 +284,9 @@ int InputString(const CKBehaviorContext &behcontext)
         if (!resetstr)
             resetstr = (char*)"";
         CKParameterOut *pout = beh->GetOutputParameter(POUT_STRINGWITHOUTCARET);
-        pout->SetValue(resetstr, strlen(resetstr) + 1);
+        pout->SetValue(resetstr, (int)strlen(resetstr) + 1);
 
-        int len = strlen(resetstr);
+        int len = (int)strlen(resetstr);
         if (len < size)
         {
             strcpy(str, resetstr);
@@ -309,16 +309,16 @@ int InputString(const CKBehaviorContext &behcontext)
         if (showcaret)
         {
             pout = beh->GetOutputParameter(POUT_STRING);
-            pout->SetValue(str, strlen(str) + 1);
+            pout->SetValue(str, (int)strlen(str) + 1);
         }
         else
         {
-            char *tmp = new char[strlen(str) + 1];
-            memcpy(tmp, str, strlen(str) + 1);
+            char *tmp = new char[(int)strlen(str) + 1];
+            memcpy(tmp, str, (int)strlen(str) + 1);
             int tcaret = caret;
-            RemoveCaret(tmp, tcaret, strlen(tmp));
+            RemoveCaret(tmp, tcaret, (int)strlen(tmp));
             pout = beh->GetOutputParameter(POUT_STRING);
-            pout->SetValue(tmp, strlen(tmp) + 1);
+            pout->SetValue(tmp, (int)strlen(tmp) + 1);
             delete[] tmp;
         }
 
@@ -347,12 +347,12 @@ int InputString(const CKBehaviorContext &behcontext)
 
             if ((multi && endkey && endkey == key) || (!multi && (key == CKKEY_RETURN || key == CKKEY_NUMPADENTER)))
             {
-                char *tmp = new char[strlen(str) + 1];
-                memcpy(tmp, str, strlen(str) + 1);
+                char *tmp = new char[(int)strlen(str) + 1];
+                memcpy(tmp, str, (int)strlen(str) + 1);
                 int tcaret = caret;
-                RemoveCaret(tmp, tcaret, strlen(tmp));
+                RemoveCaret(tmp, tcaret, (int)strlen(tmp));
                 CKParameterOut *pout = beh->GetOutputParameter(POUT_STRING);
-                pout->SetValue(tmp, strlen(tmp) + 1);
+                pout->SetValue(tmp, (int)strlen(tmp) + 1);
                 delete[] tmp;
                 beh->SetLocalParameterValue(LOCAL_POS, &caret);
                 beh->ActivateOutput(OUT_OFF, TRUE);
@@ -615,16 +615,16 @@ int InputString(const CKBehaviorContext &behcontext)
         if (showcaret)
         {
             CKParameterOut *pout = beh->GetOutputParameter(POUT_STRING);
-            pout->SetValue(str, strlen(str) + 1);
+            pout->SetValue(str, (int)strlen(str) + 1);
         }
         else
         {
-            char *tmp = new char[strlen(str) + 1];
-            memcpy(tmp, str, strlen(str) + 1);
+            char *tmp = new char[(int)strlen(str) + 1];
+            memcpy(tmp, str, (int)strlen(str) + 1);
             int tcaret = caret;
-            RemoveCaret(tmp, tcaret, strlen(tmp));
+            RemoveCaret(tmp, tcaret, (int)strlen(tmp));
             CKParameterOut *pout = beh->GetOutputParameter(POUT_STRING);
-            pout->SetValue(tmp, strlen(tmp) + 1);
+            pout->SetValue(tmp, (int)strlen(tmp) + 1);
             delete[] tmp;
         }
         beh->ActivateOutput(OUT_UPDATED, TRUE);
